Standardised API response formats and encrypted password

+++ src/main/java/com/pramurta/movie/services/PersonService.java
@@ -1,7 +1,12 @@
 package com.pramurta.movie.services;
 
+import com.pramurta.movie.domain.dtos.UpdatePersonDto;
 import com.pramurta.movie.domain.entities.Person;
+import com.pramurta.movie.repositories.MovieBookingRepository;
 import com.pramurta.movie.repositories.PersonRepository;
+import com.pramurta.movie.services.validators.PersonValidator;
+import com.pramurta.movie.services.validators.PersonValidationResult;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -9,18 +14,30 @@ import java.util.Optional;
 @Service
 public class PersonService {
     private final PersonRepository personRepository;
+    private final PersonValidator personValidator;
 
-    private static final String PERSON_NOT_FOUND_TEMPLATE = "Person with the passport number %s doesn't exist";
+    private final BCryptPasswordEncoder passwordEncoder;
 
-    public PersonService(PersonRepository personRepository) {
+    private final MovieBookingRepository movieBookingRepository;
+
+    public PersonService(PersonRepository personRepository, PersonValidator personValidator, BCryptPasswordEncoder passwordEncoder, MovieBookingRepository movieBookingRepository) {
         this.personRepository = personRepository;
+        this.personValidator = personValidator;
+        this.passwordEncoder = passwordEncoder;
+        this.movieBookingRepository = movieBookingRepository;
     }
-    public Person createPerson(Person person) {
-        if(personExists(person.getPassportNumber())) {
-            throw new RuntimeException(String.format("The person with the passport number %s already exists!"
-                    ,person.getPassportNumber()));
+    public PersonValidationResult createPerson(Person personPayload) {
+        PersonValidationResult personValidationResult = personValidator.validateCreatePerson(personPayload);
+        if(personValidationResult.getIsValid()) {
+            Person person = personValidationResult.getPerson();
+            String encryptedPassword = passwordEncoder.encode(person.getPassword());
+            person.setPassword(encryptedPassword);
+            personValidationResult.setPerson(personRepository.save(person));
         }
-        return personRepository.save(person);
+        else {
+            personValidationResult.setPerson(null);
+        }
+        return personValidationResult;
     }
 
     public Optional<Person> findPersonByPassportNumber(String passportNumber) {
@@ -31,22 +48,32 @@ public class PersonService {
         return personRepository.existsById(passportNumber);
     }
 
-    public Person updatePersonDetails(Person updatePersonPayload) {
-        if(!personExists(updatePersonPayload.getPassportNumber())) {
-            throw new RuntimeException(String.format(PERSON_NOT_FOUND_TEMPLATE,updatePersonPayload.getPassportNumber()));
+    public Person updatePersonDetails(UpdatePersonDto updatePersonPayload) throws Exception {
+        Person existingPerson = personRepository.findById(updatePersonPayload.getPassportNumber()).get();
+        PersonValidationResult personValidationResult = PersonValidationResult.builder()
+                .isValid(true)
+                .validationMessage("")
+                .build();
+        personValidator.validateUpdatePerson(updatePersonPayload, personValidationResult, existingPerson);
+        if(!personValidationResult.getIsValid()) {
+            throw new Exception(personValidationResult.getValidationMessage());
         }
-            Person existingPerson = personRepository.findById(updatePersonPayload.getPassportNumber()).get();
-            if(updatePersonPayload.getName()!=null) {
-                existingPerson.setName(updatePersonPayload.getName());
-            }
-            if(updatePersonPayload.getCardNumber()!=null) {
-                existingPerson.setCardNumber(updatePersonPayload.getCardNumber());
-            }
-            return personRepository.save(existingPerson);
-
+        if(updatePersonPayload.getName()!=null) {
+            existingPerson.setName(updatePersonPayload.getName());
+        }
+        if(updatePersonPayload.getOldPassword()!=null) {
+            existingPerson.setPassword(passwordEncoder.encode(updatePersonPayload.getNewPassword()));
+        }
+        return personRepository.save(existingPerson);
     }
-    public void removePerson(String passportNumber) {
-
+    public void removePerson(String passportNumber) throws Exception {
+        if(!personExists(passportNumber)) {
+            throw new Exception(String.format("Person with passport number: %s, doesn't exist.",passportNumber));
+        }
+        if(!movieBookingRepository.findByPassportNumber(passportNumber).isEmpty()) {
+            throw new Exception(String.format("The person with passport number: %s, has movie bookings. Delete the movie bookings first.",
+                    passportNumber));
+        }
         personRepository.deleteById(passportNumber);
     }
